+++ R/summarize_change.R
@@ -12,8 +12,9 @@
 #' or the change from baseline values (post-baseline) are then summarized.
 #'
 #' @inheritParams argument_convention
-#' @param .stats (`character`)\cr statistics to select for the table. Run `get_stats("analyze_vars_numeric)`
-#'   to see available statistics for this function.
+#' @param .stats (`character`)\cr statistics to select for the table.
+#'
+#'   Options are: ``r shQuote(get_stats("analyze_vars_numeric"))``
 #'
 #' @name summarize_change
 #' @order 1